Rename kilterboardloader file
